pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus"
version = "0.1.0"
description = "星策 — Personal AI Agent Runtime"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "aiomqtt>=2.5.1",

    # Config & env
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",

    # LLM provider (OpenAI-compatible)
    "openai>=1.30.0",

    # Data
    "pydantic>=2.0",
    "aiohttp>=3.9.0",

    # Document processing
    "pypdf>=4.0",
    "markdownify>=0.11.0",

    # Feishu / Lark SDK
    "lark-oapi>=1.4.20",
]

[project.optional-dependencies]
audio = [
    "torch>=2.3.0",
    "funasr>=1.0.0",
]
browser = [
    "playwright>=1.40.0",
]
embeddings = [
    "sentence-transformers>=2.7.0",
    "numpy>=1.26.0",
]
spreadsheet = [
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.0,<2.0",
    "ruff>=0.4.0",
]

[project.scripts]
nexus = "nexus.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["nexus"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
markers = [
    "integration: end-to-end or cross-service integration coverage",
]

[tool.ruff]
target-version = "py310"
line-length = 120
